requirements.txt
opencv-python
pydirectinput
mss
numpy
pywin32
Pillow
matplotlib
pandas